requirements-gpu.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.7
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
encodec==0.1.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
g2pkk==0.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.20.1
idna==3.6
inflect==7.0.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.0
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.58.1
numpy==1.22.0
oauthlib==3.2.2
packaging==23.2
pandas==1.5.3
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
preshed==3.0.9
protobuf==4.23.4
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pynndescent==0.5.11
pyparsing==3.1.1
pypinyin==0.50.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20230927
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torchvision==0.16.1+cu118
torchaudio==2.1.1+cu118
torch==2.1.1+cu118
tqdm==4.66.1
trainer==0.0.36
transformers==4.36.2
TTS==0.22.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.7
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
yarl==1.9.4
